support mavlink msg id greater than 255

+++ src/ui/MAVLinkDecoder.cc
@@ -75,7 +75,7 @@ void MAVLinkDecoder::receiveMessage(LinkInterface* link,mavlink_message_t messag
     Q_UNUSED(link);
     memcpy(receivedMessages+message.msgid, &message, sizeof(mavlink_message_t));
 
-    uint8_t msgid = message.msgid;
+    uint32_t msgid = message.msgid;
     const mavlink_message_info_t* msgInfo = mavlink_get_message_info(&message);
 
     // Store an arrival time for this message. This value ends up being calculated later.
@@ -220,7 +220,7 @@ void MAVLinkDecoder::emitFieldValue(mavlink_message_t* msg, int fieldid, quint64
     if (componentMulti[msg->msgid] == true) multiComponentSourceDetected = true;
 
     // Add field tree widget item
-    uint8_t msgid = msg->msgid;
+    uint32_t msgid = msg->msgid;
     if (messageFilter.contains(msgid)) return;
     QString fieldName(msgInfo->fields[fieldid].name);
     QString fieldType;
